Fix bug
Fix a bug where if a fault plane was deeper than 125 km, and surface waves were used in modelling, code was going to throw an error and stop modelling. Now, this case is recognized, so as to avoid using surface waves in modelling. The user is also given a notification.

Also, allow use of user-defined filters for surface wave data.